% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{aggr_success_rates}
\alias{aggr_success_rates}
\title{Calculating succsess rate of results}
\usage{
aggr_success_rates(aggrlist, correct)
}
\arguments{
\item{aggrlist}{list of estimate result of 'calc_bettis'}

\item{correct}{vector of correct Betti numbers}
}
\description{
Calculating succsess rate of results
}
