% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{bootstrap_homology}
\alias{bootstrap_homology}
\title{Computing cycles in subsamples}
\usage{
bootstrap_homology(X, maxdim, maxscale, const.band = 0,
  maximum.thresh = F)
}
\arguments{
\item{X}{subsamples}

\item{maxdim}{the max dimension to compute persistent homology}

\item{maxscale}{the maximum of range to compute}

\item{const.band}{constant band to discriminate cycles and noises}

\item{maximum.thresh}{use a quarter of max persistence as threshold}
}
\description{
Computing cycles in subsamples
}
