% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Making subsamples}
\usage{
bootstrapper(X, size, samples)
}
\arguments{
\item{X}{origin data}

\item{size}{the number of data points in one subsample}

\item{samples}{the number of subsamples}
}
\value{
"bootsSamples" class
}
\description{
Making subsamples
}
