% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{calc_bettis}
\alias{calc_bettis}
\title{Calculating Betti numbers of list of point cloud data}
\usage{
calc_bettis(X, maxdim, maxscale, samples, const.size = F)
}
\arguments{
\item{X}{list of point cloud data}

\item{maxdim}{the max dimension to compute persistent homology}

\item{maxscale}{the maximum of range to compute}

\item{samples}{the number of subsamples}

\item{const.size}{the number of points in one subsamples.
when 'const.size=F', the number of points in one subsamples is the 80% number of points in original data}
}
\value{
return list that have 1st and 2nd Betti numbers, the number of data, the number of points in one subsamples
  the number of subsamples, max dimension, and max scale
}
\description{
Calculating Betti numbers of list of point cloud data
}
