% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{calc_diag_centroid}
\alias{calc_diag_centroid}
\title{Calculating mean of persistence}
\usage{
calc_diag_centroid(diag = diagram)
}
\arguments{
\item{diag}{a persistence diagram}
}
\value{
mean of persistence and double mean of persistence
}
\description{
Calculating mean of persistence
}
