% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{calc_per}
\alias{calc_per}
\title{Calculating persistence of n-dimensional holes}
\usage{
calc_per(pd, dim)
}
\arguments{
\item{pd}{persisitence diagram by 'phacm'}

\item{dim}{dimension}
}
\description{
Calculating persistence of n-dimensional holes
}
