% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{calcper}
\alias{calcper}
\title{Calculating persistence of n-dimensional holes}
\usage{
calcper(pd, n)
}
\arguments{
\item{pd}{persisitence diagram by 'phacm'}

\item{n}{dimension}
}
\description{
Calculating persistence of n-dimensional holes
}
