% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{cycle_number}
\alias{cycle_number}
\title{Counting the number of data sets estimated that have n holes}
\usage{
cycle_number(aggr, dim)
}
\arguments{
\item{aggr}{estimated result using 'calc_bettis'}

\item{dim}{dimension of holes}
}
\description{
Counting the number of data sets estimated that have n holes
}
