% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figurePlot3d}
\alias{figurePlot3d}
\title{Combing 'plot3d' and 'aspect3d' to plot 3d data}
\usage{
figurePlot3d(X)
}
\arguments{
\item{X}{3d data}
}
\description{
Combing 'plot3d' and 'aspect3d' to plot 3d data
}
