% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{meetNumber}
\alias{meetNumber}
\title{Counting correct result}
\usage{
meetNumber(result, correct)
}
\arguments{
\item{result}{estimated Betti numbers}

\item{correct}{correct Betti numbers}
}
\description{
Counting correct result
}
