% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{persistence_mean}
\alias{persistence_mean}
\title{Calculating the mean of persistence except 0-persistence.}
\usage{
persistence_mean(diag)
}
\arguments{
\item{diag}{a persistence diagram.}
}
\value{
the mean of persistence.
}
\description{
Calculating the mean of persistence except 0-persistence.
}
