% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{persistence_weighted_mean}
\alias{persistence_weighted_mean}
\title{Calculate the weighted mean of persistence.
n-dimensional persistence multiplied by (the surface area of n-unit sphere/2*pi).}
\usage{
persistence_weighted_mean(diag)
}
\arguments{
\item{diag}{persistence diagram.}
}
\value{
the weighted mean of persistence.
}
\description{
Calculate the weighted mean of persistence.
n-dimensional persistence multiplied by (the surface area of n-unit sphere/2*pi).
}
