% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_circle}
\alias{plot_circle}
\title{Draw cicle in a figure with 'polygon'.}
\usage{
plot_circle(x, y, r, col)
}
\arguments{
\item{x}{x-coordinate of the center of a circle.}

\item{y}{y-coordinate of the center of a circle.}

\item{r}{the radius of a circle.}

\item{col}{color.}
}
\description{
Draw cicle in a figure with 'polygon'.
}
