% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diagrams}
\alias{plot_diagrams}
\title{Plotting persistence diagram in a list}
\usage{
plot_diagrams(diags)
}
\arguments{
\item{diags}{list of persisitence diagram}
}
\description{
Plotting persistence diagram in a list
}
