% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plotting persistence landscape}
\usage{
plot_landscape(land, dim, xlim, ylim)
}
\arguments{
\item{land}{persistence landscape calculated by 'calc_landscape'}

\item{dim}{dimension of persistence landscape to plot}

\item{xlim}{the x limits (x1, x2) of the plot of persistence landscape}

\item{ylim}{the y limits of the plot of persistence landscape}
}
\description{
Plotting persistence landscape
}
