% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{plotLandscape}
\alias{plotLandscape}
\title{Plotingpersistence  landscape}
\usage{
plotLandscape(land)
}
\arguments{
\item{land}{persistence landscape calculated by 'calc_landscape'}
}
\description{
Plotingpersistence  landscape
}
