% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rad_distribute}
\alias{rad_distribute}
\title{The function for calculate degrees of the uniform distribution respect to 3D-torus.
This function is tentative.}
\usage{
rad_distribute(r, R1, R2, theta, phi)
}
\arguments{
\item{r}{the smallest radius of 3D torus.}

\item{R1}{the largest radius of 3D torus.}

\item{R2}{the medim radius of 3D torus.}

\item{theta}{an angle of 3D torus.}

\item{phi}{an angle of 3D torus.}
}
\description{
The function for calculate degrees of the uniform distribution respect to 3D-torus.
This function is tentative.
}
