% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save2RData}
\alias{save2RData}
\title{Saving variable into RData file in "data" directory under working directory.
When there isn't "data" directory in working directory, this function make "data" directory.}
\usage{
save2RData(...)
}
\arguments{
\item{...}{one variable}
}
\description{
Saving variable into RData file in "data" directory under working directory.
When there isn't "data" directory in working directory, this function make "data" directory.
}
