% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{surface_nshpere}
\alias{surface_nshpere}
\title{Calculate surface area of a n-dimensional sphere}
\usage{
surface_nshpere(n, r = 1)
}
\arguments{
\item{n}{dimension of the sphere}

\item{r}{radius of the sphere}
}
\value{
surface area of a n-dimensional sphere
}
\description{
Calculate surface area of a n-dimensional sphere
}
