% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uniDiskMake}
\alias{uniDiskMake}
\title{Generating points distributing on disk uniformly}
\usage{
uniDiskMake(N, R)
}
\arguments{
\item{N}{the number of points}

\item{R}{radius}
}
\description{
Generating points distributing on disk uniformly
}
