% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{x3Dtorus_unif}
\alias{x3Dtorus_unif}
\title{This function samples n points from the surface or the 3D torus uniformly.
This function is tentative. There is no guarantee that points are sampled unifomly.}
\usage{
x3Dtorus_unif(n, r, R1, R2)
}
\arguments{
\item{n}{the number of points in the sample.}

\item{r}{the smallest radius of 3D torus.}

\item{R1}{the largest radius of 3D torus.}

\item{R2}{the medim radius of 3D torus.}
}
\description{
This function samples n points from the surface or the 3D torus uniformly.
This function is tentative. There is no guarantee that points are sampled unifomly.
}
