% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{xEllip_unif}
\alias{xEllip_unif}
\title{This function samples n points from the ellipsoid with equation (x/a)^2+(y/b)^2+(z/c)^2=1, uniformly with respect to its surface.
This function is tentative. There is no guarantee that points are sampled unifomly.}
\usage{
xEllip_unif(n, a, b, c)
}
\arguments{
\item{n}{the number of points in the sample.}

\item{a}{the diameter of x-axis direction.}

\item{b}{the diameter of y-axis direction.}

\item{c}{the diameter of z-axis direction.}
}
\description{
This function samples n points from the ellipsoid with equation (x/a)^2+(y/b)^2+(z/c)^2=1, uniformly with respect to its surface.
This function is tentative. There is no guarantee that points are sampled unifomly.
}
