% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{bootstrap_homology}
\alias{bootstrap_homology}
\title{Computing cycles in subsamples}
\usage{
bootstrap_homology(X, maxdim, maxscale, const.band = 0,
  maximum.thresh = F)
}
\description{
Computing cycles in subsamples
}
