% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bettis.R
\name{calc_diag_centroid}
\alias{calc_diag_centroid}
\title{Calculating the mean of persistence}
\usage{
calc_diag_centroid(diag = diagram)
}
\arguments{
\item{diag}{a persistence diagram}
}
\value{
the mean of persistence and the double mean of persistence
}
\description{
Calculating the mean of persistence
}
