% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lands}
\alias{plot_lands}
\title{Plotting persistence landscape in a list}
\usage{
plot_lands(lands, dim)
}
\arguments{
\item{lands}{list of persistence landscape}

\item{dim}{dimension to plot}
}
\description{
Plotting persistence landscape in a list
}
